perf(website): lean hydration and server config (#57)
