+++ apps/website/app/globals.css
@@ -76,128 +76,3 @@
 ::-webkit-scrollbar-thumb:hover {
   @apply bg-muted-foreground/50;
 }
-
-/* Loading animation */
-.loading-spinner {
-  @apply inline-block w-4 h-4 border-2 border-current border-t-transparent rounded-full animate-spin;
-}
-
-/* Data table styles */
-.data-table {
-  @apply w-full border-collapse;
-}
-
-.data-table th {
-  @apply bg-muted font-medium text-left p-2 border-b;
-}
-
-.data-table td {
-  @apply p-2 border-b;
-}
-
-.data-table tr:hover {
-  @apply bg-muted/50;
-}
-
-/* Custom animations and effects */
-.card-hover {
-  @apply transition-all duration-200 hover:shadow-lg hover:-translate-y-1;
-}
-
-/* Glass effect */
-.glass {
-  @apply bg-white/80 backdrop-blur-md;
-}
-
-.dark .glass {
-  @apply bg-gray-900/80;
-}
-
-/* Gradient backgrounds */
-.gradient-primary {
-  @apply bg-gradient-to-br from-primary to-accent;
-}
-
-.gradient-secondary {
-  @apply bg-gradient-to-br from-secondary to-primary/20;
-}
-
-/* Status badges */
-.badge {
-  @apply inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium;
-}
-
-.badge-success {
-  @apply bg-green-100 text-green-800 dark:bg-green-900/20 dark:text-green-400;
-}
-
-.badge-warning {
-  @apply bg-yellow-100 text-yellow-800 dark:bg-yellow-900/20 dark:text-yellow-400;
-}
-
-.badge-error {
-  @apply bg-red-100 text-red-800 dark:bg-red-900/20 dark:text-red-400;
-}
-
-.badge-info {
-  @apply bg-blue-100 text-blue-800 dark:bg-blue-900/20 dark:text-blue-400;
-}
-
-/* Print styles */
-@media print {
-  .no-print {
-    display: none !important;
-  }
-  
-  .print-break-after {
-    page-break-after: always;
-  }
-}
-
-/* Sidebar styles */
-.sidebar-fixed {
-  @apply fixed inset-y-0 z-50 flex w-72 flex-col;
-}
-
-.sidebar-content {
-  @apply flex grow flex-col gap-y-5 overflow-y-auto border-r border-gray-200 bg-white dark:bg-gray-900 dark:border-gray-800 px-6 pb-4;
-}
-
-/* Mobile menu overlay */
-.mobile-menu-overlay {
-  @apply fixed inset-0 z-50 bg-gray-900/80 lg:hidden;
-}
-
-/* Form styles */
-.form-input {
-  @apply block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-primary sm:text-sm sm:leading-6;
-}
-
-/* Button styles */
-.btn {
-  @apply inline-flex items-center justify-center rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50;
-}
-
-.btn-primary {
-  @apply bg-primary text-primary-foreground hover:bg-primary/90;
-}
-
-.btn-secondary {
-  @apply bg-secondary text-secondary-foreground hover:bg-secondary/80;
-}
-
-.btn-destructive {
-  @apply bg-destructive text-destructive-foreground hover:bg-destructive/90;
-}
-
-.btn-outline {
-  @apply border border-input bg-background hover:bg-accent hover:text-accent-foreground;
-}
-
-.btn-ghost {
-  @apply hover:bg-accent hover:text-accent-foreground;
-}
-
-.btn-link {
-  @apply text-primary underline-offset-4 hover:underline;
-}
