--- apps/website/app/providers.tsx
@@ -1,5 +1,0 @@
-'use client'
-
-export function Providers({ children }: { children: React.ReactNode }) {
-  return <>{children}</>
-}
