+++ apps/website/components/site-header/nav-data.ts
@@ -1,0 +1,12 @@
+export type NavItem = {
+  href: string
+  label: string
+}
+
+export const navItems: NavItem[] = [
+  { href: "/", label: "Home" },
+  { href: "/about", label: "About" },
+  { href: "/products", label: "Products" },
+  { href: "/ecomos", label: "EcomOS" },
+  { href: "/blog", label: "Blog" },
+]
